Catch up with previous commits

+++ DynamicDisplays/src/gov/fnal/ppd/dd/display/client/BrowserErrorListener.java
@@ -2,5 +2,5 @@ package gov.fnal.ppd.dd.display.client;
 
 public interface BrowserErrorListener {
 
-	public void errorInPageLoad(long value);
+	public void errorInPageLoad(int value);
 }
